utils/requirements.txt
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
google_cloud_bigquery==2.31.0
pip==21.1.2
wheel==0.36.2
rsa==4.8
pyasn1==0.4.8
six==1.16.0
cachetools==4.2.4
certifi==2021.10.8
urllib3==1.26.7
requests==2.26.0
packaging==21.3
python-dateutil==2.8.2
idna==3.3
pyparsing==3.0.6
setuptools==57.0.0